+++ logica.h
@@ -1,8 +1,8 @@
-double xToFramePoint(int xp, int hres, int xmax, int xmin);
-double yToFramePoint(int yp, int vres, int ymax, int ymin);
-double getL(double xw, double yw, double zw, double xe, double ye, double ze);
-double getXd(double xw, double xe, double l);
-double getYd(double yw, double ye, double l);
-double getZd(double zw, double ze, double l);
+long double xToFramePoint(int xp, int hres, int xmax, int xmin);
+long double yToFramePoint(int yp, int vres, int ymax, int ymin);
+long double getL(long double xw, long double yw, long double zw, long double xe, long double ye, long double ze);
+long double getXd(long double xw, long double xe, long double l);
+long double getYd(long double yw, long double ye, long double l);
+long double getZd(long double zw, long double ze, long double l);
